refactor(auth): move auth functions to new api module
