+++ src/components/form/Register.tsx
@@ -13,7 +13,7 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { PasswordInput } from '@/components/ui/password-input';
-import { registerUser } from '@/lib/firebase';
+import { registerUser } from '@/lib/api/auth';
 
 const FormSchema = z
   .object({
